backend/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
drf-yasg==1.20.0
python-dotenv==0.19.2
